Org: Disallows modification of accepted reservations
Previously this would lead to potentially opaque modifications of
reservations without the customer being informed of those changes.

If we end up needing to allow something like this again we should
do something more robust with workflow steps that keep the customer
in the loop.

TYPE: Bugfix
LINK: OGC-2322

+++ src/onegov/org/views/reservation.py
@@ -1111,6 +1111,16 @@ def reject_reservation(
     if view_ticket is not None and view_ticket != ticket:
         raise exc.HTTPNotFound()
 
+    if any(r.data and r.data.get('accepted') for r in targeted):
+        request.alert(_('Cannot cancel accepted reservations'))
+
+        if not request.headers.get('X-IC-Request'):
+            if view_ticket is not None:
+                return request.redirect(request.link(view_ticket))
+            return request.redirect(request.link(self))
+
+        return None
+
     client = KabaClient.from_resource(resource, request.app)
 
     # if there's a captured payment we cannot continue
@@ -1129,7 +1139,10 @@ def reject_reservation(
         return None
 
     # we need to delete the payment at the same time
-    if payment:
+    # FIXME: The price may need to be adjusted, should be handled
+    #        through the introduction of an invoice system with
+    #        individual invoice items that can be cancelled
+    if payment and len(excluded) == 0:
         request.session.delete(payment)
 
     ReservationMessage.create(targeted, ticket, request, 'rejected')
@@ -1221,6 +1234,12 @@ def reject_reservation(
 
     failed_to_revoke = False
     for reservation in targeted:
+        if payment:
+            # remove the link to the payment
+            reservation.payment = None
+            # flush, just in case, otherwise `remove_reservation` may fail
+            request.session.flush()
+
         if client and (kaba := (reservation.data or {}).get('kaba')):
             try:
                 client.revoke_visit(kaba['visit_id'])
@@ -1368,16 +1387,7 @@ def adjust_reservation(
         .filter(Reservation.id == reservation_id)
         .one_or_none()
     )
-    if reservation is None or not (
-        # is this reservation adjustable?
-        reservation.target_type == 'allocation'
-        and request.session.query(
-            reservation
-            ._target_allocations()
-            .filter(Allocation.partly_available.is_(True))
-            .exists()
-        ).scalar()
-    ) or reservation.display_start() < sedate.utcnow():
+    if reservation is None or not reservation.is_adjustable:
         if request.headers.get('X-IC-Request'):
             error = request.translate(_('Reservation not adjustable'))
             ic_data = {'message': error, 'success': False}
@@ -1451,6 +1461,9 @@ def adjust_reservation(
         if payment is None:
             pass
         elif new_reservation is not None:
+            # FIXME: The price may need to be adjusted, should be handled
+            #        through the introduction of an invoice system with
+            #        individual invoice items that can be cancelled
             new_reservation.payment = payment  # type: ignore[attr-defined]
         else:
             # restore the payment link
